pyproject.toml
[tool.poetry]
name = "eagle-eyez"
version = "11"
description = "Flashy computer vision project for mission dessert-storm, I mean dessert-line!"
authors = ["Levis J."]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.25.1"
torch = "^1.13.1"
torchaudio = "^0.13.1"
matplotlib = "^3.6.2"
torchvision = "^0.14.1"
opencv-python = "^4.7.0.68"
numpy = "^1.24.1"
black = "^22.12.0"
timm = "^0.6.12"
onnx = "^1.13.0"
ffmpeg-python = "^0.2.0"
werkzeug = "^2.2.2"
laspy = "^2.3.0"
treg = "^0.0.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.20.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
